Changed email copy to be a question
trello:
https://trello.com/c/GaBEFxnZ

We wanted the email opt in to be better and easier for the user to understand. We have iterated few options of the opt in copy heading, or adding a checkbox with “I don’t want to receive an email”.

We’ve decided these options didn’t work well enough for the interaction.

After checking with gov uk design system guide, we decided to keep the radio options and change the heading to be a simpler question, and remove unnecessary hint text.