(fix) updated questions and routes
trello: https://trello.com/c/1vDemZlk

Updated questions. One removed and One added, so it changed the ordering slightly.

And multiple children route made it harder to update.

included url mapping in .json file for easier editing of the beck and next pages, we can use it after for some sort of automation